# openapi.json
{"openapi":"3.0.2","info":{"title":"app","description":"Description","version":"0.1.0"},"paths":{"/api/v1/health_check/ping_application":{"get":{"tags":["Application Health"],"summary":"Ping Application","operationId":"ping_application_api_v1_health_check_ping_application_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PingResponse"}}}}}}},"/api/v1/health_check/ping_database":{"get":{"tags":["Application Health"],"summary":"Ping Database","operationId":"ping_database_api_v1_health_check_ping_database_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PingResponse"}}}},"500":{"description":"Database isn't working"}}}},"/api/v1/group":{"post":{"tags":["Group"],"summary":"Create Group","operationId":"create_group_api_v1_group_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupCreateRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/groups":{"get":{"tags":["Group"],"summary":"Get Groups","operationId":"get_groups_api_v1_groups_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Groups Api V1 Groups Get","type":"array","items":{"$ref":"#/components/schemas/GroupBaseScheme"}}}}}}}},"/api/v1/group/{group_id}":{"get":{"tags":["Group"],"summary":"Get Group","operationId":"get_group_api_v1_group__group_id__get","parameters":[{"required":true,"schema":{"title":"Group Id","type":"integer"},"name":"group_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupScheme"}}}},"404":{"description":"Not Found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Group"],"summary":"Update Promotion","operationId":"update_promotion_api_v1_group__group_id__put","parameters":[{"required":true,"schema":{"title":"Group Id","type":"integer"},"name":"group_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupCreateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"404":{"description":"Not Found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Group"],"summary":"Delete Promotion","operationId":"delete_promotion_api_v1_group__group_id__delete","parameters":[{"required":true,"schema":{"title":"Group Id","type":"integer"},"name":"group_id","in":"path"}],"responses":{"204":{"description":"Successful Response"},"404":{"description":"Not Found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1group/{group_id}/participant/":{"post":{"tags":["Group"],"summary":"Create Participant","operationId":"create_participant_api_v1group__group_id__participant__post","parameters":[{"required":true,"schema":{"title":"Group Id","type":"integer"},"name":"group_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParticipantCreateRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/group/{group_id}/participant/{participant_id}":{"delete":{"tags":["Group"],"summary":"Delete Member","operationId":"delete_member_api_v1_group__group_id__participant__participant_id__delete","parameters":[{"required":true,"schema":{"title":"Group Id","type":"integer"},"name":"group_id","in":"path"},{"required":true,"schema":{"title":"Participant Id","type":"integer"},"name":"participant_id","in":"path"}],"responses":{"204":{"description":"Successful Response"},"404":{"description":"Not Found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/group/{group_id}/toss":{"post":{"tags":["Group"],"summary":"Toss","operationId":"toss_api_v1_group__group_id__toss_post","parameters":[{"required":true,"schema":{"title":"Group Id","type":"integer"},"name":"group_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Toss Api V1 Group  Group Id  Toss Post","type":"array","items":{"$ref":"#/components/schemas/ParticipantScheme"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/group/{group_id}/participant/{participant_id}/recipient":{"get":{"tags":["Group"],"summary":"Get Group","operationId":"get_group_api_v1_group__group_id__participant__participant_id__recipient_get","parameters":[{"required":true,"schema":{"title":"Group Id","type":"integer"},"name":"group_id","in":"path"},{"required":true,"schema":{"title":"Participant Id","type":"integer"},"name":"participant_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParticipantBaseScheme"}}}},"404":{"description":"Not Found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"GroupBaseScheme":{"title":"GroupBaseScheme","required":["id","name"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"name":{"title":"Name","type":"string"},"description":{"title":"Description","type":"string"}}},"GroupCreateRequest":{"title":"GroupCreateRequest","required":["name"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"description":{"title":"Description","type":"string"}}},"GroupScheme":{"title":"GroupScheme","required":["id","name","participants"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"name":{"title":"Name","type":"string"},"description":{"title":"Description","type":"string"},"participants":{"title":"Participants","type":"array","items":{"$ref":"#/components/schemas/ParticipantScheme"}}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ParticipantBaseScheme":{"title":"ParticipantBaseScheme","required":["id","name"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"name":{"title":"Name","type":"string"},"wish":{"title":"Wish","type":"string"}}},"ParticipantCreateRequest":{"title":"ParticipantCreateRequest","required":["name"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"wish":{"title":"Wish","type":"string"}}},"ParticipantScheme":{"title":"ParticipantScheme","required":["id","name"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"name":{"title":"Name","type":"string"},"wish":{"title":"Wish","type":"string"},"recipient":{"$ref":"#/components/schemas/ParticipantBaseScheme"}}},"PingResponse":{"title":"PingResponse","required":["message"],"type":"object","properties":{"message":{"title":"Message","type":"string"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}},"tags":[{"name":"Application Health","description":"API health check."}]}